New post-processing functionality, fields, prefixing
This adds a new post-processing for the merger step, which applies arbitrary functions (defined in the profile) to the merged Pandas dataframe. Outlier filters have been transformed into a post-processing function and ability to define multiple functions added. There's also a new post-processing function to work on merged aggregates.

There are now several new fields with calculated ratios. Fields that are added to the document during the pre-processing phase of import are now prefixed with "dnt_" string.

In addition to this, expenses total of the red flags view should be more accurate now by taking type into account.